refactoring and renaming
We use now a new function called compute_mturns_inhomogeneous where our inhomogeneous turnover mass model is defined for both ACGs and MCGs, and it should be defined only in that function for consistency. I also changed the types of some variables from double to float, though I’m not sure it’s really necessary for all of them. In addition, I renamed mturn_a_nofb to mturn_acg_homogeneous for clarity. All tests pass

+++ src/py21cmfast/src/_functionprototypes_wrapper.h
@@ -149,6 +149,6 @@ double molecular_cooling_threshold_with_feedbacks(float z, float J_21_LW, float 
 double unconditional_hmf(double growthf, double lnM, double z, int HMF);
 double conditional_hmf(double growthf, double lnM, double delta, double sigma, int HMF);
 double expected_nhalo(double redshift);
-void compute_mturns(float z, float J_21_LW, float vcb, float Gamma12, float z_reion,
-                    double *M_turn_a, double *M_turn_m);
+void compute_mturns(double z, float J_21_LW, float vcb, float Gamma12, float z_reion,
+                    float *M_turn_a, float *M_turn_m);
 /*-----------------------*/
